[Sharing 2.0] Do not change old behaviour
In lots of places we still pass the old share array around. In order to
stay backwards compatible for now store the owner of the share in
uid_owner. And store the sharer in uid_initiator

This way the OCS endpoint can show all it wants while flat reshares are
created.